Extract cashu-wallet composition hook from useAppShellComposition
